Step 1 - I kinda forgot to upload
I forgot to upload the earlier parts of my progress, but visually not a
lot has changed. I essentially converted the earlier parts of the script
to be consistent with Gibson's 2nd edition of the book.

Outside of that, the CardProspector script was added to organize cards
into draw piles (eventually) and the Prospector script edits have moved
the cards into a Mine tableau format.

There's some sorting layer issues which I'm working on now.